bug fix: fix bug about loading glyph index zero
glyph 0 is a valid id; so it cannot and should not be used as an invalid
glyph id. A special flag is needed to indicate that the glyph slot
needs refreshing.

The glyph slot needs refreshing if one is just scaling or changing the load
flags of glyph 0, for example. Without this change, one might need to change
to glyph 1 and switch back to force glyph 0 to reload with a different
resolution/flags.

This bug also causes glyph 0 to not load if one is sequentially
loading by glyph id from zero (i.e. the first entry returned by the new
face->foreach_glyph method), as glyph is loaded only if the glyph id
differs from previous (which is zero on start up).